server/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-CORS==3.0.10
opencv-python==4.5.3.56
mediapipe==0.8.9
numpy==1.21.2
pyserial==3.5